requirements.txt
paramiko==3.0.0
psycopg2==2.9.5
SQLAlchemy==2.0.2
pyyaml
mysql-connector-python
mysqlclient